+++ regular_expressions.h
@@ -1,6 +1,8 @@
 #ifndef REGULAR_EXPRESSION_H
 #define REGULAR_EXPRESSION_H
 
+#include <stdlib.h>
+
 typedef struct type_of_regular_exp{
     int is_normal_tav;
     int is_point;
@@ -43,7 +45,7 @@ int check_if_dot_reg_exp_in_given_pattern(char* pattern);
 int check_if_circles_reg_exp_in_given_pattern(char* pattern);
 int check_if_squares_reg_exp_in_given_pattern(char *pattern);
 square_brackets find_squares_values(char *pattern);
-void set_arr_of_reg_exp_tav_round_bracket(regular_exp_tav** array_of_reg_exp_tav, char *pattern,int arr_index, int i);
-void set_arr_of_reg_exp_tav_square_bracket(regular_exp_tav** array_of_reg_exp_tav, char *pattern, int arr_index, int i);
+void set_arr_of_reg_exp_tav_round_bracket(regular_exp_tav** array_of_reg_exp_tav, char *pattern,int arr_index, size_t i);
+void set_arr_of_reg_exp_tav_square_bracket(regular_exp_tav** array_of_reg_exp_tav, char *pattern, int arr_index,size_t  i);
 
 #endif
